requirements.txt
# Core
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0

# Browser automation
selenium==4.18.1
webdriver-manager==4.0.1

# NLP
spacy==3.7.4
transformers==4.39.0
sentence-transformers==2.6.0
torch==2.2.1          # CPU version; for GPU: torch==2.2.1+cu118

# Vector search
faiss-cpu==1.8.0      # For GPU: faiss-gpu

# Database
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
alembic==1.13.1       # DB migrations

# Elasticsearch (optional)
elasticsearch==8.12.1

# Scheduling
APScheduler==3.10.4

# Data processing
numpy==1.26.4
pandas==2.2.1

# Utilities
python-dotenv==1.0.1
pydantic==2.6.3
tqdm==4.66.2
loguru==0.7.2

# Dev / testing
pytest==8.1.0
pytest-asyncio==0.23.5
